+++ index.js
@@ -9,9 +9,8 @@ import './src/color';
 import './src/gradient'; // optional gradient
 import './src/pattern.class'; // optional pattern
 import './src/shadow.class'; // optional shadow
-import './src/static_canvas.class';
-// import './src/canvas.class'; // optional interaction
-import './src/mixins/canvas_events.mixin'; // optional interaction
+import './src/canvas/static_canvas.class';
+import './src/canvas/canvas_events.mixin'; // optional interaction
 import './src/mixins/canvas_dataurl_exporter.mixin';
 import './src/mixins/canvas_serialization.mixin'; // optional serialization
 import './src/mixins/canvas_gestures.mixin'; // optional gestures
