+++ src/brushes/circle_brush.class.ts
@@ -5,7 +5,7 @@ import { Shadow } from '../shadow.class';
 import { Circle } from '../shapes/circle.class';
 import { Group } from '../shapes/group.class';
 import { getRandomInt } from '../util/internals';
-import type { Canvas } from '../mixins/canvas_events.mixin';
+import type { Canvas } from '../canvas/canvas_events';
 import { BaseBrush } from './base_brush.class';
 
 export type CircleBrushPoint = {
