+++ src/brushes/spray_brush.class.ts
@@ -4,7 +4,7 @@ import { Group } from '../shapes/group.class';
 import { Shadow } from '../shadow.class';
 import { Rect } from '../shapes/rect.class';
 import { getRandomInt } from '../util/internals';
-import type { Canvas } from '../mixins/canvas_events.mixin';
+import type { Canvas } from '../canvas/canvas_events';
 import { BaseBrush } from './base_brush.class';
 
 export type SprayBrushPoint = {
