+++ src/shapes/Object/ObjectGeometry.ts
@@ -21,8 +21,8 @@ import {
 } from '../../util/misc/matrix';
 import { degreesToRadians } from '../../util/misc/radiansDegreesConversion';
 import { sin } from '../../util/misc/sin';
-import type { Canvas } from '../../mixins/canvas_events.mixin';
-import type { StaticCanvas } from '../../static_canvas.class';
+import type { Canvas } from '../../canvas/canvas_events';
+import type { StaticCanvas } from '../../canvas/static_canvas.class';
 import { ObjectOrigin } from './ObjectOrigin';
 import { ObjectEvents } from '../../EventTypeDefs';
